feat: Refactor fee calculation and link management
- Added `cashier_common` as a dependency in `Cargo.lock` and `Cargo.toml`.
- Moved fee calculation logic to a new module `fee_calculator`.
- Introduced `calculate_link_balance_map` and `calculate_create_link_fee` functions for better fee management.
- Updated action handlers to use `source_amount` for intent total amounts.
- Refactored intent creation to include `source_amount` for clarity in fee calculations.
- Removed unused `calculator.rs` file and adjusted imports accordingly.
- Updated tests to reflect changes in fee calculation logic and intent handling.
