+++ src/cashier_backend/src/apps/link_v2/links/shared/receive_link/actions/create.rs
@@ -14,10 +14,8 @@ use cashier_backend_types::{
 };
 use cashier_common::{constant::ICP_CANISTER_PRINCIPAL, utils::convert_nat_to_u64};
 use icrc_ledger_types::icrc1::account::Account;
-use transaction_manager::{
-    intents::transfer_wallet_to_treasury::TransferWalletToTreasuryIntent,
-    utils::calculator::calculate_create_link_fee,
-};
+use fee_calculator::calculate_create_link_fee;
+use transaction_manager::intents::transfer_wallet_to_treasury::TransferWalletToTreasuryIntent;
 
 use crate::apps::link_v2::links::shared::utils::get_batch_tokens_fee_for_link;
 use uuid::Uuid;
