+++ src/cashier_backend/src/apps/link_v2/service.rs
@@ -181,16 +181,17 @@ impl<R: Repositories, M: TransactionManager + 'static> LinkV2Service<R, M> {
                     .get(&intent.id)
                     .map(Vec::as_slice)
                     .unwrap_or(&[]);
-                let fee_result = calculate_intent_fees(
+                // intent_total_amount is already set by action handler
+                if let Ok(fee_result) = calculate_intent_fees(
                     &link_model_for_fees,
                     intent,
                     transactions,
                     caller,
                     network_fee.clone(),
-                );
-                intent.intent_total_amount = Some(fee_result.intent_total_amount);
-                intent.intent_total_network_fee = Some(fee_result.intent_total_network_fee);
-                intent.intent_user_fee = Some(fee_result.intent_user_fee);
+                ) {
+                    intent.intent_total_network_fee = Some(fee_result.intent_total_network_fee);
+                    intent.intent_user_fee = Some(fee_result.intent_user_fee);
+                }
             }
         }
