+++ src/lib/fee_calculator/src/strategies/creator_to_link.rs
@@ -6,7 +6,7 @@ use cashier_backend_types::repository::intent::v2::Intent;
 use cashier_backend_types::repository::link::v1::Link;
 use cashier_backend_types::repository::transaction::v1::Transaction;
 
-use super::helpers::{calc_inbound_fee, calc_outbound_fee, get_intent_amount};
+use super::helpers::{calc_inbound_fee, calc_outbound_fee};
 use crate::traits::IntentFeeStrategy;
 use crate::types::IntentFeeResult;
 
@@ -16,12 +16,10 @@ impl IntentFeeStrategy for CreatorToLinkStrategy {
     fn calculate(
         &self,
         link: &Link,
-        intent: &Intent,
+        _intent: &Intent,
         transactions: &[Transaction],
         network_fee: Nat,
-    ) -> IntentFeeResult {
-        // intent_total_amount = amount_per_link_use_action × max_use (already set in intent)
-        let amount = get_intent_amount(intent);
+    ) -> Result<IntentFeeResult, String> {
         let max_use = link.link_use_action_max_count;
 
         // Inbound fee from transactions (source of truth)
@@ -31,11 +29,10 @@ impl IntentFeeStrategy for CreatorToLinkStrategy {
         let net_fee = inbound_fee + outbound_fee;
 
         // User pays: network fees (inbound + outbound)
-        IntentFeeResult {
-            intent_total_amount: amount,
+        Ok(IntentFeeResult {
             intent_total_network_fee: net_fee.clone(),
             intent_user_fee: net_fee,
-        }
+        })
     }
 }
 
@@ -47,14 +44,14 @@ mod tests {
     #[test]
     fn test_creator_to_link_single_use_icrc1() {
         let strategy = CreatorToLinkStrategy;
-        // intent amount = amount_per_use × max_use = 1000 × 1 = 1000
         let intent = make_intent(1000);
         let link = make_link(1);
         let txs = vec![make_icrc1_tx()];
-        let result = strategy.calculate(&link, &intent, &txs, Nat::from(10u64));
+        let result = strategy
+            .calculate(&link, &intent, &txs, Nat::from(10u64))
+            .unwrap();
 
         // inbound=10, outbound=10×1=10, net_fee=20
-        assert_eq!(result.intent_total_amount, Nat::from(1000u64));
         assert_eq!(result.intent_total_network_fee, Nat::from(20u64));
         assert_eq!(result.intent_user_fee, Nat::from(20u64));
     }
@@ -62,14 +59,14 @@ mod tests {
     #[test]
     fn test_creator_to_link_multi_use_icrc1() {
         let strategy = CreatorToLinkStrategy;
-        // intent amount = amount_per_use × max_use = 1000 × 5 = 5000
         let link = make_link(5);
         let intent = make_intent(5000);
         let txs = vec![make_icrc1_tx()];
-        let result = strategy.calculate(&link, &intent, &txs, Nat::from(10u64));
+        let result = strategy
+            .calculate(&link, &intent, &txs, Nat::from(10u64))
+            .unwrap();
 
         // inbound=10, outbound=10×5=50, net_fee=60
-        assert_eq!(result.intent_total_amount, Nat::from(5000u64));
         assert_eq!(result.intent_total_network_fee, Nat::from(60u64));
         assert_eq!(result.intent_user_fee, Nat::from(60u64));
     }
@@ -77,14 +74,14 @@ mod tests {
     #[test]
     fn test_creator_to_link_multi_use_icrc2() {
         let strategy = CreatorToLinkStrategy;
-        // intent amount = amount_per_use × max_use = 1000 × 5 = 5000
         let intent = make_intent(5000);
         let link = make_link(5);
         let txs = make_icrc2_txs();
-        let result = strategy.calculate(&link, &intent, &txs, Nat::from(10u64));
+        let result = strategy
+            .calculate(&link, &intent, &txs, Nat::from(10u64))
+            .unwrap();
 
         // inbound=20 (2 txs), outbound=10×5=50, net_fee=70
-        assert_eq!(result.intent_total_amount, Nat::from(5000u64));
         assert_eq!(result.intent_total_network_fee, Nat::from(70u64));
         assert_eq!(result.intent_user_fee, Nat::from(70u64));
     }
