+++ src/lib/fee_calculator/src/strategies/link_to_user.rs
@@ -6,7 +6,7 @@ use cashier_backend_types::repository::intent::v2::Intent;
 use cashier_backend_types::repository::link::v1::Link;
 use cashier_backend_types::repository::transaction::v1::Transaction;
 
-use super::helpers::{calc_outbound_fee, get_intent_amount};
+use super::helpers::calc_outbound_fee;
 use crate::traits::IntentFeeStrategy;
 use crate::types::IntentFeeResult;
 
@@ -16,21 +16,18 @@ impl IntentFeeStrategy for LinkToUserStrategy {
     fn calculate(
         &self,
         _link: &Link,
-        intent: &Intent,
+        _intent: &Intent,
         _transactions: &[Transaction],
         network_fee: Nat,
-    ) -> IntentFeeResult {
-        let amount = get_intent_amount(intent);
-
+    ) -> Result<IntentFeeResult, String> {
         // 0 inbound (link already has funds) + fee × 1 outbound
         let outbound_fee = calc_outbound_fee(1, &network_fee);
 
         // User receives - pays nothing
-        IntentFeeResult {
-            intent_total_amount: amount,
+        Ok(IntentFeeResult {
             intent_total_network_fee: outbound_fee,
             intent_user_fee: Nat::from(0u64),
-        }
+        })
     }
 }
 
@@ -44,10 +41,11 @@ mod tests {
         let strategy = LinkToUserStrategy;
         let intent = make_intent(1000);
         let link = make_link_default();
-        let result = strategy.calculate(&link, &intent, &[], Nat::from(10u64));
+        let result = strategy
+            .calculate(&link, &intent, &[], Nat::from(10u64))
+            .unwrap();
 
         // User receives - pays nothing
-        assert_eq!(result.intent_total_amount, Nat::from(1000u64));
         assert_eq!(result.intent_total_network_fee, Nat::from(10u64));
         assert_eq!(result.intent_user_fee, Nat::from(0u64));
     }
