+++ src/transaction_manager/src/utils/mod.rs
@@ -1,7 +1,6 @@
 // Copyright (c) 2025 Cashier Protocol Labs
 // Licensed under the MIT License (see LICENSE file in the project root)
 
-pub mod calculator;
 pub mod topological_sort;
 
 #[cfg(test)]
